Implemented train test split module
do_split is implemented which splits the data according to split dates and prediction window and finally returns the train and test data in a dict. (See comments for structure.)

created split dates in the constructor which will randomly select two dates from a month. This will be overwritten if the do_split method has split dates in it.

This commit also introduces start_time and end_time. Start_time is supposed to be the time where the first prediction will start and end_time is supposed to be the time where the last prediction will start. These are passed from the config file.